place input inside dropdownMenu
